backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
requests==2.32.3
google-genai==0.8.0
reportlab==4.2.2
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==10.4.0
pandas==2.2.2
python-dotenv==1.0.1
pydub==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
